refactor(ui): separate design tokens from cva logic and remove unused imports
Moves UI design token variants from `src/lib/variants.ts` to `src/styles/variants.ts` to enforce separation of concerns and avoid slop creep in the lib directory. Retains CVA definitions (`buttonVariants`, `badgeVariants`) in `src/lib/variants.ts`. Additionally, cleans up redundant `import { variants }` from UI components that only utilize CVA functions, reducing bundle noise.
