+++ src/layouts/Button.tsx
@@ -1,6 +1,5 @@
 import React from "react"
 import { cn } from "@/lib/utils"
-import { variants } from "@/lib/variants"
 import { buttonVariants } from "@/lib/variants"
 import { Box, BaseProps } from "./Box"
 import type { VariantProps } from "class-variance-authority"
